# bundles/SymfonyBundle-php/Model/DirectoryTreeItem.php
<?php
/**
 * DirectoryTreeItem
 *
 * PHP version 5
 *
 * @category Class
 * @package  Api\Model
 * @author   Swagger Codegen team
 * @link     https://github.com/swagger-api/swagger-codegen
 */

/**
 * Slideshow manager
 *
 * API for Slideshow manager SPA
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 *
 */

/**
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */

namespace Api\Model;

use Symfony\Component\Validator\Constraints as Assert;
use JMS\Serializer\Annotation\Type;
use JMS\Serializer\Annotation\SerializedName;

/**
 * Class representing the DirectoryTreeItem model.
 *
 * @package Api\Model
 * @author  Swagger Codegen team
 */
class DirectoryTreeItem 
{
        /**
     * @var string
     * @SerializedName("id")
     * @Assert\NotNull()
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $id;

    /**
     * @var string
     * @SerializedName("name")
     * @Assert\NotNull()
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $name;

    /**
     * @var \Api\Model\DirectoryTreeItem[]
     * @SerializedName("children")
     * @Assert\NotNull()
     * @Assert\All({
     *   @Assert\Type("Api\Model\DirectoryTreeItem")
     * })
     * @Type("array<Api\Model\DirectoryTreeItem>")
     */
    protected $children;

    /**
     * Constructor
     * @param mixed[] $data Associated array of property values initializing the model
     */
    public function __construct(array $data = null)
    {
        $this->id = isset($data['id']) ? $data['id'] : null;
        $this->name = isset($data['name']) ? $data['name'] : null;
        $this->children = isset($data['children']) ? $data['children'] : null;
    }

    /**
     * Gets id.
     *
     * @return string
     */
    public function getId()
    {
        return $this->id;
    }

    /**
     * Sets id.
     *
     * @param string $id
     *
     * @return $this
     */
    public function setId(string $id)
    {
        $this->id = $id;

        return $this;
    }

    /**
     * Gets name.
     *
     * @return string
     */
    public function getName()
    {
        return $this->name;
    }

    /**
     * Sets name.
     *
     * @param string $name
     *
     * @return $this
     */
    public function setName(string $name)
    {
        $this->name = $name;

        return $this;
    }

    /**
     * Gets children.
     *
     * @return \Api\Model\DirectoryTreeItem[]
     */
    public function getChildren()
    {
        return $this->children;
    }

    /**
     * Sets children.
     *
     * @param \Api\Model\DirectoryTreeItem[] $children
     *
     * @return $this
     */
    public function setChildren(DirectoryTreeItem $children)
    {
        $this->children = $children;

        return $this;
    }
}